+++ repository/src/main/java/com/walmartlabs/concord/repository/LastModifiedSnapshot.java
@@ -22,6 +22,8 @@ package com.walmartlabs.concord.repository;
 
 import com.walmartlabs.concord.common.FileVisitor;
 
+import java.io.IOException;
+import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.attribute.BasicFileAttributes;
 import java.nio.file.attribute.FileTime;
@@ -38,8 +40,8 @@ public class LastModifiedSnapshot implements Snapshot, FileVisitor {
     }
 
     @Override
-    public void visit(Path src, Path dst, BasicFileAttributes attrs) {
-        files.put(basePath.relativize(dst), attrs.lastModifiedTime());
+    public void visit(Path src, Path dst) throws IOException {
+        files.put(basePath.relativize(dst), Files.getLastModifiedTime(dst));
     }
 
     @Override